retains ports

+++ src/functions/isRunning.ts
@@ -4,6 +4,9 @@ import { exec } from 'child_process';
 const execSync = deasync(exec);
 
 export const isRunning = (pid: number) => {
+  if (!pid || pid === -1) {
+    return false;
+  }
   try {
     execSync(`kill -s 0 ${pid}`);
     return true;
